requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.3
panda==0.3.1
pandas==1.0.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
twilio==6.38.1
urllib3==1.25.9
Werkzeug==1.0.1
